fix: 修复大量编译错误，使应用能够编译通过
主要修复内容：
- 修复 enhanced_export_service.dart 中的参数和类型错误
- 修复 import 服务中的类型定义和方法调用问题
- 添加缺失的服务方法（encrypt/decrypt, chat, rebuildFromDate等）
- 修复 location 服务中的类型导入和方法调用
- 修复 voice_recognition_engine 添加缺失的方法
- 修复 widget 中的语法错误和类型导入
- 修复 provider 中的类型定义
- 添加存根实现以解决抽象类实例化问题
- 修复 TTS 服务类名引用（TTSService）
- 生成测试所需的 mock 文件

+++ app/lib/services/adaptive_budget_service.dart
@@ -649,7 +649,7 @@ class AdaptiveBudgetService {
     final cv = avg > 0 ? stdDev / avg : 1;
 
     // 变异系数越小，置信度越高
-    double confidence = 1 - cv.clamp(0, 1);
+    double confidence = 1 - cv.clamp(0.0, 1.0).toDouble();
 
     // 如果建议值接近历史平均，置信度更高
     final deviation = (suggested - avg).abs() / avg;
